export component, and rename tab components
