+++ packages/ui/src/lib/sidebarTabs/SidebarTabLabel.svelte
