+++ packages/ui/src/lib/sidebarTabs/SidebarTabsWrapper.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { Icon } from '@steeze-ui/svelte-icon';
 
-	import TabbedSidebarTabLabel from './TabbedSidebarTabLabel.svelte';
-	import TabbedSidebarTabList from './TabbedSidebarTabList.svelte';
+	import SidebarTabLabel from './SidebarTabLabel.svelte';
+	import SidebarTabList from './SidebarTabList.svelte';
 
 	import type { IconSource } from '@steeze-ui/heroicons/dist/types';
 	import { setContext, type ComponentType } from 'svelte';
@@ -25,18 +25,18 @@
 </script>
 
 <div class="flex">
-	<TabbedSidebarTabList>
+	<SidebarTabList>
 		{#each tabs as tab}
-			<TabbedSidebarTabLabel tabId={tab.id}>
+			<SidebarTabLabel tabId={tab.id}>
 				{#if tab.icon}
 					<Icon src={tab.icon} theme="solid" class="h-5 w-5 mb-1" aria-hidden="true" />
 				{:else if tab.rawIcon}
 					<svelte:component this={tab.rawIcon} class="h-5 w-5 mb-1" aria-hidden="true" />
 				{/if}
 				{tab.label}
-			</TabbedSidebarTabLabel>
+			</SidebarTabLabel>
 		{/each}
-	</TabbedSidebarTabList>
+	</SidebarTabList>
 
 	<div class="p-6">
 		<slot />
